# specs/001-physical-ai-textbook/contracts/api.yaml
openapi: 3.0.3
info:
  title: Physical AI Textbook API
  description: API for the Physical AI & Humanoid Robotics interactive textbook
  version: 1.0.0
  contact:
    name: Panaversity Team
    email: support@panaversity.org
servers:
  - url: https://api.physical-ai-robotics.panaversity.org/v1
    description: Production server
  - url: https://staging-api.physical-ai-robotics.panaversity.org/v1
    description: Staging server
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        background:
          $ref: '#/components/schemas/UserBackground'
        hardware:
          $ref: '#/components/schemas/UserHardware'
        createdAt:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
    UserPreferences:
      type: object
      properties:
        language:
          type: string
          enum: [en, ur]
        learningStyle:
          type: string
          enum: [visual, kinesthetic, textual, mixed]
        backgroundLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        notifications:
          type: object
          properties:
            email:
              type: boolean
            browser:
              type: boolean
    UserBackground:
      type: object
      properties:
        experience:
          type: array
          items:
            type: string
        education:
          type: string
          enum: [high-school, bachelor, master, phd, other]
        goals:
          type: array
          items:
            type: string
            enum: [research, industry, hobby, education]
    UserHardware:
      type: object
      properties:
        hasRobot:
          type: boolean
        robotType:
          type: string
        hasGPU:
          type: boolean
        platform:
          type: string
          enum: [windows, linux, macos]
...
    UserLogin:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT token for authentication
        expiresIn:
          type: number
          description: Token expiration time in seconds
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid or expired token
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...